TT-1304: Split choose_a_certified_company into separate LOA1 and LOA2 controllers
This allows a better separation of concerns and the ability to run LOA1
and LOA2 AB tests concurrently

Authors: @NasAshrafThoughtworks @callumknights
