+++ app/views/choose_a_certified_company/choose_a_certified_company_LOA1.html.erb
@@ -17,7 +17,7 @@
 
       <div id="matching-idps">
         <% @recommended_idps.each do |identity_provider| %>
-            <%= render partial: 'idp_option', locals: {recommended: true, identity_provider: identity_provider} %>
+            <%= render partial: 'choose_a_certified_company/idp_option', locals: {recommended: true, identity_provider: identity_provider} %>
         <% end %>
       </div>
