+++ app/views/choose_a_certified_company/choose_a_certified_company_LOA2.html.erb
@@ -21,7 +21,7 @@
   </div>
   <div id="matching-idps">
     <% @recommended_idps.each do |identity_provider| %>
-      <%= render partial: 'idp_option', locals: {recommended: true, identity_provider: identity_provider} %>
+      <%= render partial: 'choose_a_certified_company/idp_option', locals: {recommended: true, identity_provider: identity_provider} %>
     <% end %>
   </div>
   <% if @non_recommended_idps.any? %>
@@ -40,7 +40,7 @@
         </summary>
         <div class="panel panel-border-narrow">
           <% @non_recommended_idps.each do |identity_provider| %>
-            <%= render partial: 'idp_option', locals: {recommended: false, identity_provider: identity_provider} %>
+            <%= render partial: 'choose_a_certified_company/idp_option', locals: {recommended: false, identity_provider: identity_provider} %>
           <% end %>
         </div>
       </details>
